Add a top-level makefile to create the missing .cyclo files
This is to get rid of the warning about a pattern rule not making all
its targets. This occurs because I use a custom GCC that does not
support the -fcyclomatic-complexity option. You can turn the option off
in the STM32 IDE's compiler settings, but the autogenerated Makefiles
still contain .cyclo targets.

Reorganize the directories under Core/Src.